Move KafkaClientProvider to servicescommon
The KafkaClientProvider is used by both the master and the agent now, so it
shouldn't still be in the com.spotify.helios.agent package.
